4.21 tehty

+++ Osa 4/blogilista/controllers/blogsController.js
@@ -3,6 +3,15 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+const findUserAndValidateToken = (request, response) => {
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  return decodedToken.id
+}
+
 blogsRouter.get('/', async (request, response) => {
   //find on Mongoosen oma query, älä hämmenny
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id:1 })
@@ -11,12 +20,8 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  // eslint-disable-next-line no-undef
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+  const userId = findUserAndValidateToken(request, response)
+  const user = await User.findById(userId)
 
   const blog = new Blog({
     title: body.title,
@@ -33,6 +38,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  // eslint-disable-next-line no-undef
+  const userId = findUserAndValidateToken(request, response)
+  const user = await User.findById(userId)
+  if ( blog.user._id.toString() !== user._id.toString() ){
+    return response.status(403).json({ error: 'blog can only be deleted by the person who originally added it' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
